src/requirements.txt
python-dotenv~=1.0.1
pyodbc==5.1.0
streamlit==1.38.0
streamlit-chat==0.1.1
langchain==0.2.10
langchain-openai==0.1.17
langchain-community==0.2.9
graphrag==0.2.0
psycopg2-binary==2.9.9
future==1.0.0
pandas==2.2.2
tiktoken==0.7.0
requests==2.32.3
openai==1.43.0
